Changed the Google Log in
I changed a few things from the old one, I tried to add a bit more commenting to give clarity as to what is what. In doing so I added a user class in hopes that anything necessary - personID, email, token etc could be stored there in case you'd like but obviously if you have a better and preferred method its no problem.
Added a push notification system, was not able to test it thoroughly as I had said in the group chat, please give it a go, it should be able to push to a specific device via its token or a broadcast based on the subject. I believe I left token specific messaging. This is in the fire base messaging service class. 
I added a vague logo-esque image and some text to the login activity as well.
I have attempted to make sure that  any specific additions that are new are shown as new segments.

Need to still do some unit tests, if there are any changes or work done to it that improves or you have done to modify it please just send upload the code and tell me with a - Look here.